+++ dramatiq/__init__.py
@@ -1,12 +1,15 @@
 from .actor import Actor, actor  # noqa
 from .broker import Broker, Consumer, MessageProxy, get_broker, set_broker  # noqa
+from .encoder import Encoder, JSONEncoder, PickleEncoder  # noqa
 from .errors import (  # noqa
-    DramatiqError, BrokerError, ActorNotFound, QueueNotFound, ConnectionError,
-    ConnectionClosed, ConnectionFailed, RateLimitExceeded
+    DramatiqError,
+    BrokerError, ActorNotFound, QueueNotFound,
+    ConnectionError, ConnectionClosed, ConnectionFailed,
+    RateLimitExceeded,
 )
 from .generic import GenericActor  # noqa
 from .logging import get_logger  # noqa
-from .message import Message  # noqa
+from .message import Message, get_encoder, set_encoder  # noqa
 from .middleware import Middleware  # noqa
 from .worker import Worker  # noqa